// src/hocphan/dto/hocphan.dto.ts
// create-hocphan.dto.ts

export class CreateHocphanDto {
    mahocphan: string;
    tenhp: string;
    thutu: number;
    nam: string;
    hocki: number;
    diadiemhoc: string;
    soluong: number;
    gvphutrach: string;
    tutiet: string;
    dentiet: string;
    mota: string;

    motamonhoc?: string;
    tenmonhoc: string;
    mamonhoc: string;
    sotc: number;
}